create shoulder subsystem and commands

+++ src/main/java/frc/robot/utils/Gains.java
@@ -5,10 +5,10 @@ public class Gains {
 	public final double kI;
 	public final double kD;
 	public final double kF;
-	public final int kIzone;
+	public final double kIzone;
 	public final double kPeakOutput;
 	
-	public Gains(double _kP, double _kI, double _kD, double _kF, int _kIzone, double _kPeakOutput){
+	public Gains(double _kP, double _kI, double _kD, double _kF, double _kIzone, double _kPeakOutput){
 		kP = _kP;
 		kI = _kI;
 		kD = _kD;
